backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.11.7
python-dotenv==1.0.1
httpx==0.28.1
bcrypt==4.1.2
PyJWT==2.8.0

